requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
anyio=3.7.0=pypi_0
bzip2=1.0.8=h620ffc9_4
ca-certificates=2023.01.10=hca03da5_0
certifi=2023.5.7=pypi_0
cffi=1.15.1=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
cryptography=41.0.1=pypi_0
fastapi=0.96.0=pypi_0
h11=0.14.0=pypi_0
idna=3.4=pypi_0
libffi=3.4.4=hca03da5_0
msal=1.22.0=pypi_0
ncurses=6.4=h313beb8_0
openssl=1.1.1t=h1a28f6b_0
pip=23.0.1=py311hca03da5_0
pycparser=2.21=pypi_0
pydantic=1.10.8=pypi_0
pyjwt=2.7.0=pypi_0
python=3.11.3=hc0d8a6c_0
python-dotenv=1.0.0=pypi_0
readline=8.2=h1a28f6b_0
requests=2.31.0=pypi_0
setuptools=67.8.0=py311hca03da5_0
sniffio=1.3.0=pypi_0
sqlite=3.41.2=h80987f9_0
starlette=0.27.0=pypi_0
tk=8.6.12=hb8d0fd4_0
typing-extensions=4.6.3=pypi_0
tzdata=2023c=h04d1e81_0
urllib3=2.0.2=pypi_0
uvicorn=0.22.0=pypi_0
wheel=0.38.4=py311hca03da5_0
xz=5.4.2=h80987f9_0
zlib=1.2.13=h5a0b063_0
